tests/requirements.txt
requests>=2.31.0
azure-identity>=1.15.0
azure-mgmt-datafactory>=6.0.0
azure-mgmt-synapse>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
